requirements.txt
streamlit
requests
faiss-cpu
numpy
pandas
textblob
sentence-transformers
matplotlib
wordcloud
seaborn
